+++ smartsim/entity/entity.py
@@ -29,11 +29,11 @@ from __future__ import annotations
 import abc
 import typing as t
 
-from smartsim.launchable.jobGroup import JobGroup
+from smartsim.launchable.job_group import JobGroup
 
 if t.TYPE_CHECKING:
     from smartsim.launchable.job import Job
-    from smartsim.settings.launchSettings import LaunchSettings
+    from smartsim.settings.launch_settings import LaunchSettings
 
 
 class TelemetryConfiguration:
