+++ smartsim/launchable/colocated_job_group.py
@@ -29,21 +29,15 @@ from __future__ import annotations
 import typing as t
 from copy import deepcopy
 
-from smartsim.log import get_logger
-
-from .._core.utils.helpers import check_name
-from .basejob import BaseJob
-from .baseJobGroup import BaseJobGroup
-
-logger = get_logger(__name__)
+from .base_job import BaseJob
+from .base_job_group import BaseJobGroup
 
 if t.TYPE_CHECKING:
     from typing_extensions import Self
 
 
-@t.final
-class JobGroup(BaseJobGroup):
-    """A job group holds references to multiple jobs that
+class ColocatedJobGroup(BaseJobGroup):
+    """A colocated job group holds references to multiple jobs that
     will be executed all at the same time when resources
     permit. Execution is blocked until resources are available.
     """
@@ -51,24 +45,9 @@ class JobGroup(BaseJobGroup):
     def __init__(
         self,
         jobs: t.List[BaseJob],
-        name: str = "job_group",
     ) -> None:
         super().__init__()
         self._jobs = deepcopy(jobs)
-        self._name = name
-        check_name(self._name)
-
-    @property
-    def name(self) -> str:
-        """Retrieves the name of the JobGroup."""
-        return self._name
-
-    @name.setter
-    def name(self, name: str) -> None:
-        """Sets the name of the JobGroup."""
-        check_name(name)
-        logger.debug(f'Overwriting Job name from "{self._name}" to "{name}"')
-        self._name = name
 
     @property
     def jobs(self) -> t.List[BaseJob]:
@@ -91,6 +70,6 @@ class JobGroup(BaseJobGroup):
 
     def __str__(self) -> str:  # pragma: no-cover
         """Returns a string representation of the collection of
-        job groups.
+        colocated job groups.
         """
-        return f"Job Groups: {self.jobs}"
+        return f"Colocated Jobs: {self.jobs}"
