+++ smartsim/launchable/job.py
@@ -32,7 +32,7 @@ from copy import deepcopy
 
 from smartsim._core.commands.launch_commands import LaunchCommands
 from smartsim._core.utils.helpers import check_name
-from smartsim.launchable.basejob import BaseJob
+from smartsim.launchable.base_job import BaseJob
 from smartsim.log import get_logger
 from smartsim.settings import LaunchSettings
