+++ smartsim/settings/arguments/__init__.py
@@ -24,7 +24,7 @@
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-from .batchArguments import BatchArguments
-from .launchArguments import LaunchArguments
+from .batch_arguments import BatchArguments
+from .launch_arguments import LaunchArguments
 
 __all__ = ["LaunchArguments", "BatchArguments"]
