+++ tests/temp_tests/test_core/test_commands/test_launchCommands.py
@@ -29,7 +29,7 @@ import pytest
 from smartsim._core.commands.command import Command
 from smartsim._core.commands.command_list import CommandList
 from smartsim._core.commands.launch_commands import LaunchCommands
-from smartsim.settings.launchCommand import LauncherType
+from smartsim.settings.launch_command import LauncherType
 
 pytestmark = pytest.mark.group_a
