+++ tests/temp_tests/test_settings/test_lsfScheduler.py
@@ -26,7 +26,7 @@
 import pytest
 
 from smartsim.settings import BatchSettings
-from smartsim.settings.batchCommand import SchedulerType
+from smartsim.settings.batch_command import SchedulerType
 
 pytestmark = pytest.mark.group_a
